+++ src/amd/OclCLI.cpp
@@ -188,8 +188,8 @@ int OclCLI::getHints(const GpuContext &ctx, xmrig::Config *config) const
         hints |= CNv2;
     }
 
-    if (config->algorithm().algo() == xmrig::CRYPTONIGHT_ULTRALITE) {
-        hints |= Ultralite;
+    if (config->algorithm().algo() == xmrig::CRYPTONIGHT_FEMTO) {
+        hints |= Femto;
     }
 
     if (config->algorithm().algo() == xmrig::CRYPTONIGHT_EXTREMELITE) {
@@ -252,7 +252,7 @@ void OclCLI::parse(std::vector<int> &vector, const char *arg) const
 
 size_t OclCLI::getMaxThreads(const GpuContext &ctx, xmrig::Algo algo, int hints)
 {
-    const size_t ratio = (algo == xmrig::CRYPTONIGHT_LITE || algo == xmrig::CRYPTONIGHT_ULTRALITE || algo == xmrig::CRYPTONIGHT_EXTREMELITE) ? 2u : 1u;
+    const size_t ratio = (algo == xmrig::CRYPTONIGHT_LITE || algo == xmrig::CRYPTONIGHT_FEMTO || algo == xmrig::CRYPTONIGHT_EXTREMELITE) ? 2u : 1u;
     if (ctx.vendor == xmrig::OCL_VENDOR_INTEL) {
         return ratio * ctx.computeUnits * 8;
     }
@@ -291,7 +291,7 @@ size_t OclCLI::getPossibleIntensity(const GpuContext &ctx, size_t maxThreads, si
 size_t OclCLI::worksizeByHints(int hints)
 {
     if (hints & Vega) {
-        if (hints & Ultralite) {
+        if (hints & Femto) {
             return 64;
         }