+++ src/crypto/CryptoNight_x86.h
@@ -661,10 +661,10 @@ extern xmrig::CpuThread::cn_mainloop_fun        cn_half_mainloop_ryzen_asm;
 extern xmrig::CpuThread::cn_mainloop_fun        cn_half_mainloop_bulldozer_asm;
 extern xmrig::CpuThread::cn_mainloop_double_fun cn_half_double_mainloop_sandybridge_asm;
 
-extern xmrig::CpuThread::cn_mainloop_fun        cn_trtl_mainloop_ivybridge_asm;
-extern xmrig::CpuThread::cn_mainloop_fun        cn_trtl_mainloop_ryzen_asm;
-extern xmrig::CpuThread::cn_mainloop_fun        cn_trtl_mainloop_bulldozer_asm;
-extern xmrig::CpuThread::cn_mainloop_double_fun cn_trtl_double_mainloop_sandybridge_asm;
+extern xmrig::CpuThread::cn_mainloop_fun        cn_cat_mainloop_ivybridge_asm;
+extern xmrig::CpuThread::cn_mainloop_fun        cn_cat_mainloop_ryzen_asm;
+extern xmrig::CpuThread::cn_mainloop_fun        cn_cat_mainloop_bulldozer_asm;
+extern xmrig::CpuThread::cn_mainloop_double_fun cn_cat_double_mainloop_sandybridge_asm;
 
 void wow_compile_code(const V4_Instruction* code, int code_size, void* machine_code, xmrig::Assembly ASM);
 void v4_compile_code(const V4_Instruction* code, int code_size, void* machine_code, xmrig::Assembly ASM);
@@ -735,13 +735,13 @@ inline void cryptonight_single_hash_asm(const uint8_t *__restrict__ input, size_
     }
     else if (VARIANT == xmrig::VARIANT_TURTLE) {
         if (ASM == xmrig::ASM_INTEL) {
-            cn_trtl_mainloop_ivybridge_asm(ctx[0]);
+            cn_cat_mainloop_ivybridge_asm(ctx[0]);
         }
         else if (ASM == xmrig::ASM_RYZEN) {
-            cn_trtl_mainloop_ryzen_asm(ctx[0]);
+            cn_cat_mainloop_ryzen_asm(ctx[0]);
         }
         else {
-            cn_trtl_mainloop_bulldozer_asm(ctx[0]);
+            cn_cat_mainloop_bulldozer_asm(ctx[0]);
         }
     }
     else if (xmrig::cn_is_cryptonight_r<VARIANT>()) {
@@ -780,7 +780,7 @@ inline void cryptonight_double_hash_asm(const uint8_t *__restrict__ input, size_
         cn_half_double_mainloop_sandybridge_asm(ctx[0], ctx[1]);
     }
     else if (VARIANT == xmrig::VARIANT_TURTLE) {
-        cn_trtl_double_mainloop_sandybridge_asm(ctx[0], ctx[1]);
+        cn_cat_double_mainloop_sandybridge_asm(ctx[0], ctx[1]);
     }
     else if (xmrig::cn_is_cryptonight_r<VARIANT>()) {
         ctx[0]->generated_code_double(ctx[0], ctx[1]);