+++ src/net/strategies/DonateStrategy.cpp
@@ -59,7 +59,7 @@ xmrig::DonateStrategy::DonateStrategy(int level, const char *user, Algo algo, IS
         m_pools.push_back(Pool("donate2.graef.in", 8443, userId, nullptr, true, false, true));
     } else if (algo == xmrig::Algo::CRYPTONIGHT_LITE) {
         m_pools.push_back(Pool("donate2.graef.in", 1080, userId, nullptr, true, false, true));
-    } else if (algo == xmrig::Algo::CRYPTONIGHT_ULTRALITE) {
+    } else if (algo == xmrig::Algo::CRYPTONIGHT_FEMTO) {
         m_pools.push_back(Pool("donate2.graef.in", 8090, userId, nullptr, true, false, true));
     } else if (algo == xmrig::Algo::CRYPTONIGHT_EXTREMELITE) {
         m_pools.push_back(Pool("donate2.graef.in", 9091, userId, nullptr, true, false, true));
@@ -71,7 +71,7 @@ xmrig::DonateStrategy::DonateStrategy(int level, const char *user, Algo algo, IS
         m_pools.push_back(Pool("donate2.graef.in", 9000, userId, nullptr, false, false, true));
     } else if (algo == xmrig::Algo::CRYPTONIGHT_LITE) {
         m_pools.push_back(Pool("donate2.graef.in", 7000, userId, nullptr, false, false, true));
-    } else if (algo == xmrig::Algo::CRYPTONIGHT_ULTRALITE) {
+    } else if (algo == xmrig::Algo::CRYPTONIGHT_FEMTO) {
         m_pools.push_back(Pool("donate2.graef.in", 8088, userId, nullptr, false, false, true));
     } else if (algo == xmrig::Algo::CRYPTONIGHT_EXTREMELITE) {
         m_pools.push_back(Pool("donate2.graef.in", 8188, userId, nullptr, false, false, true));